% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfBindingSites.R
\docType{methods}
\name{getChipSeqHits,tfBindingSites-method}
\alias{getChipSeqHits,tfBindingSites-method}
\alias{getChipSeqHits}
\title{get a table of ChIP-seq hits of the specified tf, in the specified region}
\usage{
\S4method{getChipSeqHits}{tfBindingSites}(obj, chromosome, min, max)
}
\arguments{
\item{obj}{An object of class tfBindingSites}

\item{chromosome}{character}

\item{min}{numeric}

\item{max}{numeric}
}
\value{
A data.frame
}
\description{
get a table of ChIP-seq hits of the specified tf, in the specified region
}
