% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfBindingSites.R
\docType{methods}
\name{getFimoHits,tfBindingSites-method}
\alias{getFimoHits,tfBindingSites-method}
\alias{getFimoHits}
\title{get a table of FIMO hits in the regions specified in the incoming data.frame}
\usage{
\S4method{getFimoHits}{tfBindingSites}(obj, tbl.regions, pvalThreshold)
}
\arguments{
\item{obj}{An object of class tfBindingSites}

\item{tbl}{a data.frame with (at least) chr, start, end columns}
}
\value{
a data.frame
}
\description{
get a table of FIMO hits in the regions specified in the incoming data.frame
}
