% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfBindingSites.R
\docType{methods}
\name{getMotifMatcherHits,tfBindingSites-method}
\alias{getMotifMatcherHits,tfBindingSites-method}
\alias{getMotifMatcherHits}
\title{get a table of bioconductor (MotifMatcher) hits in the regions specified in the incoming data.frame}
\usage{
\S4method{getMotifMatcherHits}{tfBindingSites}(obj, tbl.regions, pfms,
  matchThreshold = 0.7)
}
\arguments{
\item{obj}{An object of class tfBindingSites}

\item{tbl.regions}{a data.frame with (at least) chr, start, end columns}
}
\value{
a data.frame
}
\description{
get a table of bioconductor (MotifMatcher) hits in the regions specified in the incoming data.frame
}
