% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfBindingSites.R
\name{tfBindingSitesCtor}
\alias{tfBindingSitesCtor}
\title{Create a tfBindingSitesCtor object}
\usage{
tfBindingSitesCtor(tf, quiet = TRUE)
}
\arguments{
\item{tf}{character, a transcription factor gene symbol name}

\item{quiet}{do or do not print progress information}
}
\value{
An object of the tfBindingSites class
}
\description{
once configured, this object reports motif match to supplied (ChIP-seq) sequences - recommended
}
